Initial creation
This site has all of the most resent blog post entries. There is a list
of things to do:  1. Create a template for the project pages 2. Create a
listing page just for project pages 3. Clean up the post page template
navigation 4. Place an excerpt of the latest post on the main page 5.
Update the about page to give better information about the base of the
template and the hyde project 6. Place contact information in about page
7. create .htaccess file to make url host_url/2010/10/something.html go
to host_url/blog/2010/10/something… or the other way around (need
research)